requirements.txt
numpy
torch
opencv-python
pillow
tqdm
scipy
